Skip WithInstanceManagerServiceTrait
